requirements.txt
cryptography==1.5.2
Django==1.10.1
django-webpack-loader==0.3.3
flake8==3.0.4
Pillow==3.3.1
pycountry==1.20
python-dateutil==2.5.3
pytz==2016.6.1
Scrapy==1.1.3
scrapy-djangoitem==1.1.1
